+++ NEWS.md
@@ -3,6 +3,7 @@
 * `dropDuplicateIDs()` allows dropping duplicate IDs based on number of missings on selected variables (#67)
 * `changeValLabels()` and `changeMissings()` allow changing value labels and missing tags for multiple variables at once (#33)
 * `recodeGADS()` allows recoding multiple variables at once (#107)
+* `merge.GADSdat()` allows automatically assigning NAs created through merging a missing code and a value label (#1)
 
 ## bug fixes
 * `extractData2()` and `extractData()` no longer throw an error if multiple values of the same variable are labeled `NA` (#96)